+++ src/components/sign-up-breadcrumb.js
@@ -48,6 +48,7 @@ const styles = EStyleSheet.create({
     marginHorizontal: 4,
   },
   text: {
+    fontFamily: "roboto",
     fontSize: "0.6rem",
     marginTop: 10,
     alignSelf: "center",
@@ -79,17 +80,23 @@ const SignUpBreadCrumb = props => {
           <MaterialIcons name="face" size={iconSize} color={defaultColor} style={[styles.icon, highlightId === 1 && styles.highlight]} />
           <Text style={[styles.text, highlightId === 1 && styles.textHighlight]}>{locale.identification}</Text>
         </View>
+
         <IonIcons name="ios-arrow-round-forward" size={iconSize} style={styles.arrowIcon} color={defaultColor} />
+
         <View style={styles.iconContainer}>
           <MaterialIcons name="person" size={iconSize} color={defaultColor} style={[styles.icon, highlightId === 2 && styles.highlight]} />
           <Text style={[styles.text, highlightId === 2 && styles.textHighlight]}>{locale.profileData}</Text>
         </View>
+
         <IonIcons name="ios-arrow-round-forward" size={iconSize} style={styles.arrowIcon} color={defaultColor} />
+
         <View style={styles.iconContainer}>
           <MaterialCommunityIcons name="pin" size={iconSize} color={defaultColor} style={[styles.icon, highlightId === 3 && styles.highlight]} />
           <Text style={[styles.text, highlightId === 3 && styles.textHighlight]}>{locale.locality}</Text>
         </View>
+
         <IonIcons name="ios-arrow-round-forward" size={iconSize} style={styles.arrowIcon} color={defaultColor} />
+
         <View style={styles.iconContainer}>
           <MaterialCommunityIcons name="check-outline" size={iconSize} color={defaultColor} style={[styles.icon, highlightId === 4 && styles.highlight]} />
           <Text style={[styles.text, highlightId === 4 && styles.textHighlight]}>{locale.conclude}</Text>